backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Marshmallow==0.14.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.3.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
alembic==1.11.1
openai==0.27.8
requests==2.31.0
PyJWT==2.7.0
python-multipart==0.0.6
marshmallow==3.20.1  # 锁定兼容版本